isdn: i4l: Send all ISDN traffic to pppd daemon
This patch sends _all_ ISDN traffic to the pppd daemon, not only the various
control protocol packets. This allows to let the PPP subsystem (ppp_generic)
handle (de)compression without requiring to fiddle with I4L-specific IOCTLs.
The downside is that (de)compression protocols specific to I4L (LZS, to be
exact) are not available.

Signed-off-by: Christoph Schulz <fli4l@kristov.de>

[Converted to Git patch by Alexander Dahl <post@lespocky.de>]